Add tests for the TagService
Add some coverage for the TagService. Modifies the retrieve method
from accepting a Tag object to accepting a "name" to retrieve a
tag for consistency with the user branch.

Note: There is a class PDOTestHelper inside the test case for the
TagService which should be removed once the user branch is merged
into master.